requirements.txt
certifi==2025.11.12
charset-normalizer==3.4.4
idna==3.11
psycopg2-binary==2.9.11
PyYAML==6.0.3
requests==2.32.5
urllib3==2.6.2
websocket-client==1.9.0
